Refactored test helpers into an own module
